Agent: enhance systemd service dependency
The create-cluster-and-infraenv.service "Requires" assisted-service.service, this will make it be deactived when assisted-service.service fail to start. When assisted-service.service restarts, sinece create-cluster-and-infraenv.service was deactived, it will not restart automatically. Therefore, it is better to use "Wants=", it is a weaker version of Requires, and the dependency failure will not affect the dependent. https://www.freedesktop.org/software/systemd/man/systemd.unit.html#Requires=